refactor: replace models/keys/address with lumos hd
